window stuff, change funky.json to test

+++ Source/World.cpp
@@ -202,7 +202,7 @@ void World::Tick() {
     auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(end_time - tpsStartTime).count();
 
     if (duration >= 1000.0) {
-        ticksPerSecond = (float)totalTicks / (duration / 1000.0);
+        ticksPerSecond = static_cast<unsigned int>(static_cast<float>(totalTicks) / (duration / 1000.0));
         totalTicks = 0;
         tpsStartTime = end_time;
     }
